% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-ete.R
\name{get_answers_ete}
\alias{get_answers_ete}
\title{Réponses à la vague loisirs estivaux}
\usage{
get_answers_ete(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs estivaux}
}
\value{
Tibble avec une ligne par réponse et par participant·e
}
\description{
Cette fonction renvoie les réponses de tout·es les participant·es à la vague loisirs estivaux
}
