% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-hiver.R
\name{get_answers_hiver}
\alias{get_answers_hiver}
\title{Réponses à la vague loisirs hivernaux}
\usage{
get_answers_hiver(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs hivernaux}
}
\value{
Tibble avec une ligne par réponse et par participant·e
}
\description{
Cette fonction renvoie les réponses de tout·es les participant·es à la vague loisirs hivernaux
}
