% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-PDCn.R
\name{get_answers_pdcn}
\alias{get_answers_pdcn}
\title{Réponses à la vague PDCn}
\usage{
get_answers_pdcn(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague PDCn}
}
\value{
Tibble avec une ligne par réponse et par participant·e
}
\description{
Cette fonction renvoie les réponses de tout·es les participant·es à la vague PDCn
}
