% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rythme.R
\name{get_answers_rythme}
\alias{get_answers_rythme}
\title{Réponses à la vague rythme}
\usage{
get_answers_rythme(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague rythme}
}
\value{
Tibble avec une ligne par réponse et par participant·e
}
\description{
Cette fonction renvoie les réponses de tout·es les participant·es à la vague rythme
}
