% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave2.R
\name{get_answers_wave2}
\alias{get_answers_wave2}
\title{Réponses à la vague 2}
\usage{
get_answers_wave2(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 2}
}
\value{
Tibble avec une ligne par réponse et par participant·e
}
\description{
Cette fonction renvoie les réponses de tout·es les participant·es à la vague 2
}
