% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave3.R
\name{get_answers_wave3}
\alias{get_answers_wave3}
\title{Réponses à la vague 3}
\usage{
get_answers_wave3(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 3}
}
\value{
Tibble avec une ligne par réponse et par participant·e
}
\description{
Cette fonction renvoie les réponses de tout·es les participant·es à la vague 3
}
