% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-ete.R
\name{get_documentation_ete}
\alias{get_documentation_ete}
\title{Documentation de la vague loisirs estivaux}
\usage{
get_documentation_ete(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs estivaux}
}
\value{
Tibble avec la description de chaque variable
}
\description{
Cette fonction permet d'obtenir une documentation de toutes les variables de la vague sur les loisirs estivaux
}
