% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-hiver.R
\name{get_documentation_hiver}
\alias{get_documentation_hiver}
\title{Documentation de la vague loisirs hivernaux}
\usage{
get_documentation_hiver(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs hivernaux}
}
\value{
Tibble avec la description de chaque variable
}
\description{
Cette fonction permet d'obtenir une documentation de toutes les variables de la vague sur les loisirs hivernaux
}
