% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-PDCn.R
\name{get_documentation_pdcn}
\alias{get_documentation_pdcn}
\title{Documentation de la vague PDCn}
\usage{
get_documentation_pdcn(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague PDCn}
}
\value{
Tibble avec la description de chaque variable
}
\description{
Cette fonction permet d'obtenir une documentation de toutes les variables de la vague PDCn
}
