% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rythme.R
\name{get_documentation_rythme}
\alias{get_documentation_rythme}
\title{Documentation de la vague rythme}
\usage{
get_documentation_rythme(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague rythme}
}
\value{
Tibble avec la description de chaque variable
}
\description{
Cette fonction permet d'obtenir une documentation de toutes les variables de la vague rythme
}
