% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave3.R
\name{get_documentation_wave3}
\alias{get_documentation_wave3}
\title{Documentation de la vague 3}
\usage{
get_documentation_wave3(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 3}
}
\value{
Tibble avec la description de chaque variable
}
\description{
Cette fonction permet d'obtenir une documentation de toutes les variables de la vague 3
}
