% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_labels}
\alias{get_labels}
\title{Extract the labels from labelled columns in a data frame}
\usage{
get_labels(x, ..., names_to = "variable_name", name = "name", value = "value")
}
\arguments{
\item{x}{Data frame}

\item{...}{Arguments passed to methods.}

\item{names_to}{String, specifies the name of the column, in the output data frame, storing the
names of labelled columns of x}

\item{name}{String, specifies the name of the column \code{tibble::enframe} uses to store the
name.}

\item{value}{String, specifies the name of the column \code{tibble::enframe} uses to store the
value.}
}
\value{
Data frame
}
\description{
\code{get_labels} creates a data frame with one row per question variable and labelled value.
}
