% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-ete.R
\name{get_participants_ete}
\alias{get_participants_ete}
\title{Participant·es de la vague loisirs estivaux}
\usage{
get_participants_ete(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs estivaux}
}
\value{
Tibble avec les caractèristiques de chaque participant·e
}
\description{
Cette fonction retourne les caractèristiques des participant·es de la vague loisirs estivaux
}
