% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-hiver.R
\name{get_participants_hiver}
\alias{get_participants_hiver}
\title{Participant·es de la vague loisirs hivernaux}
\usage{
get_participants_hiver(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs hivernaux}
}
\value{
Tibble avec les caractèristiques de chaque participant·e
}
\description{
Cette fonction retourne les caractèristiques des participant·es de la vague loisirs hivernaux
}
