% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-PDCn.R
\name{get_participants_pdcn}
\alias{get_participants_pdcn}
\title{Participant·es de la vague PDCn}
\usage{
get_participants_pdcn(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague PDCn}
}
\value{
Tibble avec les caractèristiques de chaque participant·e
}
\description{
Cette fonction retourne les caractèristiques des participant·es de la vague PDCn
}
