% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rythme.R
\name{get_participants_rythme}
\alias{get_participants_rythme}
\title{Participant·es de la vague rythme}
\usage{
get_participants_rythme(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague rythme}
}
\value{
Tibble avec les caractèristiques de chaque participant·e
}
\description{
Cette fonction retourne les caractèristiques des participant·es de la vague rythme
}
