% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave1.R
\name{get_participants_wave1}
\alias{get_participants_wave1}
\title{Participant·es de la vague 1}
\usage{
get_participants_wave1(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 1}
}
\value{
Tibble avec les caractèristiques de chaque participant·e
}
\description{
Cette fonction retourne les caractèristiques des participant·es de la vague 1
}
