% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave2.R
\name{get_participants_wave2}
\alias{get_participants_wave2}
\title{Participant·es de la vague 2}
\usage{
get_participants_wave2(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 2}
}
\value{
Tibble avec les caractèristiques de chaque participant·e
}
\description{
Cette fonction retourne les caractèristiques des participant·es de la vague 2
}
