% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave3.R
\name{get_participants_wave3}
\alias{get_participants_wave3}
\title{Participant·es de la vague 3}
\usage{
get_participants_wave3(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 3}
}
\value{
Tibble avec les caractèristiques de chaque participant·e
}
\description{
Cette fonction retourne les caractèristiques des participant·es de la vague 3
}
