% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-ete.R
\name{get_questions_ete}
\alias{get_questions_ete}
\title{Questions de la vague loisirs estivaux}
\usage{
get_questions_ete(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs estivaux}
}
\value{
Tibble avec la description de chaque question
}
\description{
Cette fonction permet d'obtenir l'ensemble des questions posées aux participant·es de 
la vague sur les loisirs estivaux ainsi que la section du questionnaire à laquelle
elles appartiennent
}
