% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-hiver.R
\name{get_questions_hiver}
\alias{get_questions_hiver}
\title{Questions de la vague loisirs hivernaux}
\usage{
get_questions_hiver(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs hivernaux}
}
\value{
Tibble avec la description de chaque question
}
\description{
Cette fonction permet d'obtenir l'ensemble des questions posées aux participant·es de 
la vague sur les loisirs hivernaux ainsi que la section du questionnaire à laquelle
elles appartiennent
}
