% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-PDCn.R
\name{get_questions_pdcn}
\alias{get_questions_pdcn}
\title{Questions de la vague PDCn}
\usage{
get_questions_pdcn(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague PDCn}
}
\value{
Tibble avec la description de chaque question
}
\description{
Cette fonction permet d'obtenir l'ensemble des questions posées aux participant·es de 
la vague PDCn ainsi que la section du questionnaire à laquelle
elles appartiennent
}
