% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rythme.R
\name{get_questions_rythme}
\alias{get_questions_rythme}
\title{Questions de la vague rythme}
\usage{
get_questions_rythme(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague rythme}
}
\value{
Tibble avec la description de chaque question
}
\description{
Cette fonction permet d'obtenir l'ensemble des questions posées aux participant·es de 
la vague rythme ainsi que la section du questionnaire à laquelle
elles appartiennent
}
