% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave1.R
\name{get_questions_wave1}
\alias{get_questions_wave1}
\title{Questions de la vague 1}
\usage{
get_questions_wave1(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 1}
}
\value{
Tibble avec la description de chaque question
}
\description{
Cette fonction permet d'obtenir l'ensemble des questions posées aux participant·es de 
la vague 1 ainsi que la section du questionnaire à laquelle
elles appartiennent
}
