% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave2.R
\name{get_questions_wave2}
\alias{get_questions_wave2}
\title{Questions de la vague 2}
\usage{
get_questions_wave2(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 2}
}
\value{
Tibble avec la description de chaque question
}
\description{
Cette fonction permet d'obtenir l'ensemble des questions posées aux participant·es de 
la vague 2 ainsi que la section du questionnaire à laquelle
elles appartiennent
}
