% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave3.R
\name{get_questions_wave3}
\alias{get_questions_wave3}
\title{Questions de la vague 3}
\usage{
get_questions_wave3(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 3}
}
\value{
Tibble avec la description de chaque question
}
\description{
Cette fonction permet d'obtenir l'ensemble des questions posées aux participant·es de 
la vague 3 ainsi que la section du questionnaire à laquelle
elles appartiennent
}
