% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-ete.R
\name{get_section_ete}
\alias{get_section_ete}
\title{Section de la vague loisirs estivaux}
\usage{
get_section_ete()
}
\value{
Liste avec le nom des sections
}
\description{
Cette fonction liste les sections du questionnaire de la vague loisirs estivaux
}
