% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-hiver.R
\name{get_section_hiver}
\alias{get_section_hiver}
\title{Section de la vague loisirs hivernaux}
\usage{
get_section_hiver()
}
\value{
Liste avec le nom des sections
}
\description{
Cette fonction liste les sections du questionnaire de la vague loisirs hivernaux
}
