% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-PDCn.R
\name{get_section_pdcn}
\alias{get_section_pdcn}
\title{Section de la vague PDCn}
\usage{
get_section_pdcn()
}
\value{
Liste avec le nom des sections
}
\description{
Cette fonction liste les sections du questionnaire de la vague PDCn
}
