% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rythme.R
\name{get_section_rythme}
\alias{get_section_rythme}
\title{Section de la vague rythme}
\usage{
get_section_rythme()
}
\value{
Liste avec le nom des sections
}
\description{
Cette fonction liste les sections du questionnaire de la vague rythme
}
