% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave1.R
\name{get_section_wave1}
\alias{get_section_wave1}
\title{Section de la vague 1}
\usage{
get_section_wave1()
}
\value{
Liste avec le nom des sections
}
\description{
Cette fonction liste les sections du questionnaire de la vague 1
}
