% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave2.R
\name{get_section_wave2}
\alias{get_section_wave2}
\title{Section de la vague 2}
\usage{
get_section_wave2(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 2}
}
\value{
Liste avec le nom des sections
}
\description{
Cette fonction liste les sections du questionnaire de la vague 2
}
