% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave3.R
\name{get_section_wave3}
\alias{get_section_wave3}
\title{Section de la vague 3}
\usage{
get_section_wave3()
}
\value{
Liste avec le nom des sections
}
\description{
Cette fonction liste les sections du questionnaire de la vague 3
}
