% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-ete.R
\name{get_survey_completion_ete}
\alias{get_survey_completion_ete}
\title{Métadonnées de la vague loisirs estivaux}
\usage{
get_survey_completion_ete(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs estivaux}
}
\value{
Tibble avec les métadonnées
}
\description{
Cette fonction retourne les métadonnées de la vague sur les loisirs estivaux
}
