% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loisir-hiver.R
\name{get_survey_completion_hiver}
\alias{get_survey_completion_hiver}
\title{Métadonnées de la vague loisirs hivernaux}
\usage{
get_survey_completion_hiver(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague sur les loisirs hivernaux}
}
\value{
Tibble avec les métadonnées
}
\description{
Cette fonction retourne les métadonnées de la vague sur les loisirs hivernaux
}
