% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-PDCn.R
\name{get_survey_completion_pdcn}
\alias{get_survey_completion_pdcn}
\title{Métadonnées de la vague PDCn}
\usage{
get_survey_completion_pdcn(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague PDCn}
}
\value{
Tibble avec les métadonnées
}
\description{
Cette fonction retourne les métadonnées de la vague PDCn
}
