% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rythme.R
\name{get_survey_completion_rythme}
\alias{get_survey_completion_rythme}
\title{Métadonnées de la vague rythme}
\usage{
get_survey_completion_rythme(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague rythme}
}
\value{
Tibble avec les métadonnées
}
\description{
Cette fonction retourne les métadonnées de la vague rythme
}
