% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave2.R
\name{get_survey_completion_wave2}
\alias{get_survey_completion_wave2}
\title{Métadonnées de la vague 2}
\usage{
get_survey_completion_wave2(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 2}
}
\value{
Tibble avec les métadonnées
}
\description{
Cette fonction retourne les métadonnées de la vague 2
}
