% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-wave3.R
\name{get_survey_completion_wave3}
\alias{get_survey_completion_wave3}
\title{Métadonnées de la vague 3}
\usage{
get_survey_completion_wave3(data)
}
\arguments{
\item{data}{Tibble avec les données de la vague 3}
}
\value{
Tibble avec les métadonnées
}
\description{
Cette fonction retourne les métadonnées de la vague 3
}
