% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labels_to_string}
\alias{labels_to_string}
\title{Creates a string from a question's labels}
\usage{
labels_to_string(data, index)
}
\arguments{
\item{data}{tibble?, SPSS dataset}

\item{index}{int, index of the question}
}
\value{
string
}
\description{
Creates a string from a question's labels
}
