% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_responses}
\alias{pivot_responses}
\title{Pivot responses data frame to longer format}
\usage{
pivot_responses(
  x,
  ...,
  id_column = "participant_code",
  selection_type = c("character", "numeric"),
  remove_nas = TRUE,
  names_to = "name",
  values_to = "value"
)
}
\arguments{
\item{x}{Data frame}

\item{...}{Arguments passed to methods.}

\item{id_column}{String, specifies the column containing the sample id}

\item{selection_type}{String, specifies whether character or numeric columns should be
pivoted.}

\item{remove_nas}{Boolean, specifies whether rows with NA should be removed}

\item{names_to}{String, specifies the name of the column in which the question_codes will be
stored. Default is "name".}

\item{values_to}{String, specifies the name of the column in which the response text or value
will be stored. Default is "value".}
}
\value{
Data frame, with the specified escape characters removed from character columns
}
\description{
\code{pivot_responses} pivots the responses data frame to a longer format, with one row per
sample x question.
}
