% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_escapeseqs}
\alias{remove_escapeseqs}
\title{Remove specific escape sequences}
\usage{
remove_escapeseqs(x, ..., characters = c("\\n", "\\t"))
}
\arguments{
\item{x}{Data frame}

\item{...}{Arguments passed to methods.}

\item{characters}{Character vector, specifies which escape characters should be removed from
character columns of x.}
}
\value{
Data frame, with the specified escape characters removed from character columns
}
\description{
\code{remove_escapeseqs} allows the user to remove escape sequences from the text. This is
particularly useful when \n or \t have been inserted, leading to issues when saving to .tsv.
}
