% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_whitespace}
\alias{remove_whitespace}
\title{removes whitespaces at the begining of a string}
\usage{
remove_whitespace(x)
}
\arguments{
\item{x}{string}
}
\description{
removes whitespaces at the begining of a string
}
