% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zap_all}
\alias{zap_all}
\title{Apply all zaps on data frame}
\usage{
zap_all(x)
}
\arguments{
\item{x}{Data frame}
}
\value{
Data frame
}
\description{
\code{zap_all} applies all \code{haven::zap_*} functions except zap_empty to the data frame
passed as argument.
}
